+++ Domain/VBMS.Domain/Enums/VillageGroupMemberStatus.cs
@@ -1,6 +1,6 @@
 ﻿namespace VBMS.Domain.Enums;
 
-public enum VillageGroupMemberStatus : byte
+public enum VillageGroupMemberStatus : int
 {
     Inactive = 0,
     Active = 1,
